+++ api/src/pcapi/alembic/versions/20210831_4317a4dc233e_custom_reimbursement_rule.py
@@ -4,7 +4,7 @@ import sqlalchemy as sa
 from sqlalchemy import func
 from sqlalchemy.dialects import postgresql
 
-from pcapi.core.payments.models import CustomReimbursementRule
+from pcapi.core.finance.models import CustomReimbursementRule
 
 
 # revision identifiers, used by Alembic.
