+++ api/src/pcapi/core/finance/exceptions.py
@@ -16,3 +16,21 @@ class WrongLengthSiret(InvalidSiret):
 
 class NotAllDigitSiret(InvalidSiret):
     pass
+
+
+class ReimbursementRuleValidationError(Exception):
+    pass
+
+
+class ConflictingReimbursementRule(ReimbursementRuleValidationError):
+    def __init__(self, msg, conflicts):  # type: ignore [no-untyped-def]
+        super().__init__(msg)
+        self.conflicts = conflicts
+
+
+class WrongDateForReimbursementRule(ReimbursementRuleValidationError):
+    pass
+
+
+class UnknownSubcategoryForReimbursementRule(ReimbursementRuleValidationError):
+    pass
