+++ api/src/pcapi/core/finance/repository.py
@@ -11,7 +11,6 @@ import pcapi.core.bookings.models as bookings_models
 import pcapi.core.educational.models as educational_models
 import pcapi.core.offerers.models as offerers_models
 import pcapi.core.offers.models as offers_models
-import pcapi.core.payments.models as payments_models
 import pcapi.core.users.models as users_models
 from pcapi.models import db
 from pcapi.models.feature import FeatureToggle
@@ -162,9 +161,9 @@ def has_active_or_future_custom_reimbursement_rule(offer: offers_models.Offer) -
     """
     now = datetime.datetime.utcnow()
     timespan = db_utils.make_timerange(start=now, end=None)
-    query = payments_models.CustomReimbursementRule.query.filter(
-        payments_models.CustomReimbursementRule.offerId == offer.id,
-        payments_models.CustomReimbursementRule.timespan.overlaps(timespan),
+    query = models.CustomReimbursementRule.query.filter(
+        models.CustomReimbursementRule.offerId == offer.id,
+        models.CustomReimbursementRule.timespan.overlaps(timespan),
     ).exists()
     return db.session.query(query).scalar()
