+++ api/src/pcapi/domain/reimbursement.py
@@ -5,16 +5,16 @@ from decimal import Decimal
 from pcapi.core.bookings.models import Booking
 from pcapi.core.categories import subcategories
 from pcapi.core.educational.models import CollectiveBooking
-from pcapi.core.finance import conf as finance_conf
+import pcapi.core.finance.conf as finance_conf
+import pcapi.core.finance.models as finance_models
 from pcapi.core.offers.models import Offer
-import pcapi.core.payments.models as payments_models
 
 
 # A new set rules are in effect as of 1 September 2021 (i.e. 31 August 22:00 UTC)
 SEPTEMBER_2021 = datetime.datetime(2021, 9, 1) - datetime.timedelta(hours=2)
 
 
-class DigitalThingsReimbursement(payments_models.ReimbursementRule):
+class DigitalThingsReimbursement(finance_models.ReimbursementRule):
     rate = Decimal(0)
     description = "Pas de remboursement pour les offres digitales"
     group = finance_conf.RuleGroups.NOT_REIMBURSED  # type: ignore [assignment]
@@ -26,7 +26,7 @@ class DigitalThingsReimbursement(payments_models.ReimbursementRule):
         return offer.subcategory.reimbursement_rule == subcategories.ReimbursementRuleChoices.NOT_REIMBURSED.value
 
 
-class EducationalOffersReimbursement(payments_models.ReimbursementRule):
+class EducationalOffersReimbursement(finance_models.ReimbursementRule):
     rate = Decimal(1)
     description = "Remboursement total pour les offres éducationnelles"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -40,7 +40,7 @@ class EducationalOffersReimbursement(payments_models.ReimbursementRule):
         return Decimal(booking.collectiveStock.price) * self.rate
 
 
-class PhysicalOffersReimbursement(payments_models.ReimbursementRule):
+class PhysicalOffersReimbursement(finance_models.ReimbursementRule):
     rate = Decimal(1)
     description = "Remboursement total pour les offres physiques"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -51,7 +51,7 @@ class PhysicalOffersReimbursement(payments_models.ReimbursementRule):
         return is_relevant_for_standard_reimbursement_rule(booking.stock.offer)
 
 
-class MaxReimbursementByOfferer(payments_models.ReimbursementRule):
+class MaxReimbursementByOfferer(finance_models.ReimbursementRule):
     # This rule is not used anymore.
     rate = Decimal(0)
     description = "Pas de remboursement au dessus du plafond de 20 000 € par acteur culturel"
@@ -65,7 +65,7 @@ class MaxReimbursementByOfferer(payments_models.ReimbursementRule):
         return cumulative_revenue > 20000
 
 
-class LegacyPreSeptember2021ReimbursementRateByVenueBetween20000And40000(payments_models.ReimbursementRule):
+class LegacyPreSeptember2021ReimbursementRateByVenueBetween20000And40000(finance_models.ReimbursementRule):
     rate = Decimal("0.95")
     description = "Remboursement à 95% entre 20 000 € et 40 000 € par lieu"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -78,7 +78,7 @@ class LegacyPreSeptember2021ReimbursementRateByVenueBetween20000And40000(payment
         return 20000 < cumulative_revenue <= 40000
 
 
-class LegacyPreSeptember2021ReimbursementRateByVenueBetween40000And150000(payments_models.ReimbursementRule):
+class LegacyPreSeptember2021ReimbursementRateByVenueBetween40000And150000(finance_models.ReimbursementRule):
     rate = Decimal("0.85")
     description = "Remboursement à 85% entre 40 000 € et 150 000 € par lieu"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -91,7 +91,7 @@ class LegacyPreSeptember2021ReimbursementRateByVenueBetween40000And150000(paymen
         return 40000 < cumulative_revenue <= 150000
 
 
-class LegacyPreSeptember2021ReimbursementRateByVenueAbove150000(payments_models.ReimbursementRule):
+class LegacyPreSeptember2021ReimbursementRateByVenueAbove150000(finance_models.ReimbursementRule):
     rate = Decimal("0.70")
     description = "Remboursement à 70% au dessus de 150 000 € par lieu"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -104,7 +104,7 @@ class LegacyPreSeptember2021ReimbursementRateByVenueAbove150000(payments_models.
         return cumulative_revenue > 150000
 
 
-class ReimbursementRateByVenueBetween20000And40000(payments_models.ReimbursementRule):
+class ReimbursementRateByVenueBetween20000And40000(finance_models.ReimbursementRule):
     rate = Decimal("0.95")
     description = "Remboursement à 95% entre 20 000 € et 40 000 € par lieu (>= 2021-09-01)"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -117,7 +117,7 @@ class ReimbursementRateByVenueBetween20000And40000(payments_models.Reimbursement
         return 20000 < cumulative_revenue <= 40000
 
 
-class ReimbursementRateByVenueBetween40000And150000(payments_models.ReimbursementRule):
+class ReimbursementRateByVenueBetween40000And150000(finance_models.ReimbursementRule):
     rate = Decimal("0.92")
     description = "Remboursement à 92% entre 40 000 € et 150 000 € par lieu (>= 2021-09-01)"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -130,7 +130,7 @@ class ReimbursementRateByVenueBetween40000And150000(payments_models.Reimbursemen
         return 40000 < cumulative_revenue <= 150000
 
 
-class ReimbursementRateByVenueAbove150000(payments_models.ReimbursementRule):
+class ReimbursementRateByVenueAbove150000(finance_models.ReimbursementRule):
     rate = Decimal("0.90")
     description = "Remboursement à 90% au dessus de 150 000 € par lieu (>= 2021-09-01)"
     group = finance_conf.RuleGroups.STANDARD  # type: ignore [assignment]
@@ -143,7 +143,7 @@ class ReimbursementRateByVenueAbove150000(payments_models.ReimbursementRule):
         return cumulative_revenue > 150000
 
 
-class ReimbursementRateForBookBelow20000(payments_models.ReimbursementRule):
+class ReimbursementRateForBookBelow20000(finance_models.ReimbursementRule):
     rate = Decimal(1)
     description = "Remboursement à 100% jusqu'à 20 000 € pour les livres"
     group = finance_conf.RuleGroups.BOOK  # type: ignore [assignment]
@@ -156,7 +156,7 @@ class ReimbursementRateForBookBelow20000(payments_models.ReimbursementRule):
         return cumulative_revenue <= 20000
 
 
-class ReimbursementRateForBookAbove20000(payments_models.ReimbursementRule):
+class ReimbursementRateForBookAbove20000(finance_models.ReimbursementRule):
     rate = Decimal("0.95")
     description = "Remboursement à 95% au dessus de 20 000 € pour les livres"
     group = finance_conf.RuleGroups.BOOK  # type: ignore [assignment]
@@ -202,13 +202,13 @@ assert [r.description for r in REGULAR_RULES] == [
 @dataclass
 class BookingReimbursement:
     booking: Booking
-    rule: payments_models.ReimbursementRule
+    rule: finance_models.ReimbursementRule
     reimbursed_amount: Decimal
 
 
 class CustomRuleFinder:
     def __init__(self):  # type: ignore [no-untyped-def]
-        self.rules = payments_models.CustomReimbursementRule.query.all()
+        self.rules = finance_models.CustomReimbursementRule.query.all()
         self.rules_by_offer = self._partition_by_field("offerId")
         self.rules_by_offerer = self._partition_by_field("offererId")
 
@@ -218,7 +218,7 @@ class CustomRuleFinder:
             cache.setdefault(getattr(rule, field), []).append(rule)
         return cache
 
-    def get_rule(self, booking: Booking) -> payments_models.CustomReimbursementRule | None:
+    def get_rule(self, booking: Booking) -> finance_models.CustomReimbursementRule | None:
         for rule in self.rules_by_offer.get(booking.stock.offerId, ()):
             if rule.matches(booking):
                 return rule
@@ -230,7 +230,7 @@ class CustomRuleFinder:
 
 def get_reimbursement_rule(
     booking: Booking | CollectiveBooking, custom_rule_finder: CustomRuleFinder, cumulative_revenue: Decimal
-) -> payments_models.ReimbursementRule:
+) -> finance_models.ReimbursementRule:
     if isinstance(booking, CollectiveBooking):
         return EducationalOffersReimbursement()
