+++ api/tests/core/finance/test_repository.py
@@ -13,7 +13,6 @@ from pcapi.core.finance.models import BankInformationStatus
 import pcapi.core.offerers.factories as offerers_factories
 import pcapi.core.offerers.models as offerers_models
 import pcapi.core.offers.factories as offers_factories
-import pcapi.core.payments.factories as payments_factories
 from pcapi.core.testing import override_features
 import pcapi.core.users.factories as users_factories
 from pcapi.models import db
@@ -424,21 +423,21 @@ class HasActiveOrFutureCustomRemibursementRuleTest:
     def test_active_rule(self):
         now = datetime.datetime.utcnow()
         timespan = (now - datetime.timedelta(days=1), now + datetime.timedelta(days=1))
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
+        rule = factories.CustomReimbursementRuleFactory(timespan=timespan)
         offer = rule.offer
         assert repository.has_active_or_future_custom_reimbursement_rule(offer)
 
     def test_future_rule(self):
         now = datetime.datetime.utcnow()
         timespan = (now + datetime.timedelta(days=1), None)
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
+        rule = factories.CustomReimbursementRuleFactory(timespan=timespan)
         offer = rule.offer
         assert repository.has_active_or_future_custom_reimbursement_rule(offer)
 
     def test_past_rule(self):
         now = datetime.datetime.utcnow()
         timespan = (now - datetime.timedelta(days=2), now - datetime.timedelta(days=1))
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
+        rule = factories.CustomReimbursementRuleFactory(timespan=timespan)
         offer = rule.offer
         assert not repository.has_active_or_future_custom_reimbursement_rule(offer)
