+++ api/tests/core/payments/test_api.py
@@ -1,26 +1,20 @@
 from datetime import datetime
 from datetime import time
-from datetime import timedelta
 from decimal import Decimal
 
 from dateutil.relativedelta import relativedelta
 from freezegun import freeze_time
 import pytest
-import pytz
 
 from pcapi.core.fraud import factories as fraud_factories
 from pcapi.core.fraud import models as fraud_models
-import pcapi.core.offerers.factories as offerers_factories
-import pcapi.core.offers.factories as offers_factories
 from pcapi.core.payments import api
 from pcapi.core.payments import exceptions
-from pcapi.core.payments import factories as payments_factories
 from pcapi.core.payments.models import Deposit
 from pcapi.core.payments.models import DepositType
 from pcapi.core.testing import override_settings
 from pcapi.core.users import factories as users_factories
 from pcapi.core.users import models as users_models
-from pcapi.models import db
 
 
 pytestmark = pytest.mark.usefixtures("db_session")
@@ -102,81 +96,3 @@ class CreateDepositTest:
         # Then
         assert Deposit.query.filter(Deposit.userId == beneficiary.id).count() == 1
         assert error.value.errors["user"] == ['Cet utilisateur a déjà été crédité de la subvention "GRANT_18".']
-
-
-class CreateOffererReimbursementRuleTest:
-    @freeze_time("2021-10-01 00:00:00")
-    def test_create_rule(self):
-        offerer = offerers_factories.OffererFactory()
-        start = pytz.utc.localize(datetime.today() + timedelta(days=1))
-        end = pytz.utc.localize(datetime.today() + timedelta(days=2))
-        rule = api.create_offerer_reimbursement_rule(
-            offerer.id, subcategories=["VOD"], rate=0.8, start_date=start, end_date=end
-        )
-
-        db.session.refresh(rule)
-        assert rule.offerer == offerer
-        assert rule.subcategories == ["VOD"]
-        assert rule.rate == Decimal("0.8")
-        assert rule.timespan.lower == datetime(2021, 10, 2, 0, 0)
-        assert rule.timespan.upper == datetime(2021, 10, 3, 0, 0)
-
-    def test_validation(self):
-        # Validation is thoroughly verified in `test_validation.py`.
-        # This is just an integration test.
-        offerer = offerers_factories.OffererFactory()
-        start = (datetime.today() + timedelta(days=1)).astimezone(pytz.utc)
-        with pytest.raises(exceptions.UnknownSubcategoryForReimbursementRule):
-            api.create_offerer_reimbursement_rule(offerer.id, subcategories=["UNKNOWN"], rate=0.8, start_date=start)
-
-
-class CreateOfferReimbursementRuleTest:
-    @freeze_time("2021-10-01 00:00:00")
-    def test_create_rule(self):
-        offer = offers_factories.OfferFactory()
-        start = pytz.utc.localize(datetime.today() + timedelta(days=1))
-        end = pytz.utc.localize(datetime.today() + timedelta(days=2))
-        rule = api.create_offer_reimbursement_rule(offer.id, amount=12.34, start_date=start, end_date=end)
-
-        db.session.refresh(rule)
-        assert rule.offer == offer
-        assert rule.amount == Decimal("12.34")
-        assert rule.timespan.lower == datetime(2021, 10, 2, 0, 0)
-        assert rule.timespan.upper == datetime(2021, 10, 3, 0, 0)
-
-    def test_validation(self):
-        # Validation is thoroughly verified in `test_validation.py`.
-        # This is just an integration test.
-        offer = offers_factories.OfferFactory()
-        start = (datetime.today() + timedelta(days=1)).astimezone(pytz.utc)
-        payments_factories.CustomReimbursementRuleFactory(offer=offer, timespan=(start, None))
-        with pytest.raises(exceptions.ConflictingReimbursementRule):
-            api.create_offer_reimbursement_rule(offer.id, amount=12.34, start_date=start)
-
-
-class EditReimbursementRuleTest:
-    def test_edit_rule(self):
-        timespan = (pytz.utc.localize(datetime(2021, 1, 1)), None)
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
-        end = pytz.utc.localize(datetime(2030, 10, 3, 0, 0))
-        api.edit_reimbursement_rule(rule, end_date=end)
-
-        db.session.refresh(rule)
-        assert rule.timespan.lower == datetime(2021, 1, 1, 0, 0)  # unchanged
-        assert rule.timespan.upper == datetime(2030, 10, 3, 0, 0)
-
-    def test_cannot_change_existing_end_date(self):
-        today = datetime.today()
-        timespan = (today - timedelta(days=10), today)
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
-        with pytest.raises(exceptions.WrongDateForReimbursementRule):
-            api.edit_reimbursement_rule(rule, end_date=today + timedelta(days=5))
-
-    def test_validation(self):
-        # Validation is thoroughly verified in `test_validation.py`.
-        # This is just an integration test.
-        timespan = (datetime.today() - timedelta(days=10), None)
-        rule = payments_factories.CustomReimbursementRuleFactory(timespan=timespan)
-        end = pytz.utc.localize(datetime.today())
-        with pytest.raises(exceptions.WrongDateForReimbursementRule):
-            api.edit_reimbursement_rule(rule, end_date=end)
